cleanup/restructured
