--- src/cli/domain/package-server-script/utils/index.js
@@ -1,11 +1,0 @@
-'use strict';
-
-var compress = require('./compress');
-var wrapLoops = require('./wrapLoops');
-var missingDependencies = require('./missingDependencies');
-
-module.exports = {
-  compress: compress,
-  wrapLoops: wrapLoops,
-  missingDependencies: missingDependencies
-};
